discord_bot/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py==2.3.2
python-dotenv==1.0.1
asyncpg==0.22.0
